migrate: re-run programmatic migrations on error
This commit modifies the migration framework to re-attempt programmatic
migrations if they error during a migration, on the next run.
Previously, if a programmatic migration failed, but their associated SQL
migration succeeded, the database version would be set to a dirty state,
and require manual intervention in order to reset the SQL migration and
re-attempt it + the programmatic migration.

The new re-attempt mechanism is achieved by ensuring that a migration
can only be either an SQL migration or a programmatic migration, but not
both. This way, if a programmatic migration errors, the database version
will be reset to the previous version prior to executing the
programmatic migration, and the programmatic migration will be
re-attempted on the next run.