migrate: rename PostStepCallback
Rename the PostStepCallback to ProgrammaticMigration. In the upcoming
commits, we will redefine the defintion of a migration, to either be an
SQL migration, or a Programmatic GoLang based migration.

As the migration therefore will not be executed **post** the SQL based
migration anymore, we rename it to better reflect its purpose, i.e. that
a migration is either only an SQL based migration, or a Programmatic
one.