migrate: make rerun of failed programmatic migs optional
The last commit introduced functionality to rewind the database to the
prior version before the migration was attempted, when a programmatic
migration fails. This is done to ensure that such migrations are re-run
on the next startup and to ensure that the database is not left in a
dirty state.

However, to support the old behavior that did not re-run failed
programmatic migrations, we add a new flag called RerunOnError which is
part of the migration definition.
When this flag is set to false, a failing programmatic migration will
leave the database dirty at the migration version of the failed
programmatic migration, which matches the old behavior.